Adjusted Pak and hook system
