+++ UnrealEngineModLoader/Hooks.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace Hooks
+{
+	void SetupHooks();
+};
