+++ UnrealEngineModLoader/UnrealEngineModLoader.sln
@@ -9,24 +9,14 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UnrealEngineModLauncher", "
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DE3BC2AD-2683-4D76-AD72-95D98164E9C7}.Debug|x64.ActiveCfg = Debug|x64
-		{DE3BC2AD-2683-4D76-AD72-95D98164E9C7}.Debug|x64.Build.0 = Debug|x64
-		{DE3BC2AD-2683-4D76-AD72-95D98164E9C7}.Debug|x86.ActiveCfg = Debug|Win32
-		{DE3BC2AD-2683-4D76-AD72-95D98164E9C7}.Debug|x86.Build.0 = Debug|Win32
 		{DE3BC2AD-2683-4D76-AD72-95D98164E9C7}.Release|x64.ActiveCfg = Release|x64
 		{DE3BC2AD-2683-4D76-AD72-95D98164E9C7}.Release|x64.Build.0 = Release|x64
 		{DE3BC2AD-2683-4D76-AD72-95D98164E9C7}.Release|x86.ActiveCfg = Release|Win32
 		{DE3BC2AD-2683-4D76-AD72-95D98164E9C7}.Release|x86.Build.0 = Release|Win32
-		{5BF2489F-778A-4781-8666-853F6685931E}.Debug|x64.ActiveCfg = Debug|x64
-		{5BF2489F-778A-4781-8666-853F6685931E}.Debug|x64.Build.0 = Debug|x64
-		{5BF2489F-778A-4781-8666-853F6685931E}.Debug|x86.ActiveCfg = Debug|Win32
-		{5BF2489F-778A-4781-8666-853F6685931E}.Debug|x86.Build.0 = Debug|Win32
 		{5BF2489F-778A-4781-8666-853F6685931E}.Release|x64.ActiveCfg = Release|x64
 		{5BF2489F-778A-4781-8666-853F6685931E}.Release|x64.Build.0 = Release|x64
 		{5BF2489F-778A-4781-8666-853F6685931E}.Release|x86.ActiveCfg = Release|Win32
